FIX - Code duplication by create component auth code
